requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
environs==11.0.0
exceptiongroup==1.2.1
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
marshmallow==3.21.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
python-slugify==8.0.4
slugify==0.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
text-unidecode==1.3
typing_extensions==4.11.0
uvicorn==0.29.0
